Enable generation of the Constants.ts file from live site code, to ease maintenance as the site updates
Update Constants.ts to the latest

Harden Packet.aboutModel and fix some bugs with it. It is far less black
magic now and more accurate, if somewhat incomplete. It works for all the
packets we typically merge and use .aboutModel for though. And the
previous code was wrong in that it would never identify the correct model
for packets about models who have been around for a long long time. For
instance, the model with user id 119 is still active and the code would
have never identified her packets.

Add an interface definition for the sMessage format that comes with ROOMDATA packets