Better reconnect logic when the servers are unresponsive to the initial connect request.
In the last day or two, chat servers have been failing to respond
to the initial connect request. Node appears to have an extremely
long timeout on the initial connect and worse, the reconnect logic
that did exist was entirely tired to connect finishing.

This change factors out the reconnect logic and extends our "zombie
socket" detection to work when the zombie is waiting to connect.