Support WebSocket connections to MFC in addition to traditional TCP sockets
Create a WebSocket based Client by passing true as the third parameter
on the constructor. Here's an example of logging in as a guest via
websockets:
    `new Client("guest", "guest", true)`

By default, this value is false and all clients will continue to use
the traditional TCP sockets that MFC's Flash-based chat connections
have always used.

This flag only changes the underlying connection type. All other
aspects of MFCAuto should work exactly as they always have.

I'm adding this for two reasons:
    1. Flash is being slowly deprecated, and I won't be surprised if some day soon MFC drops support for it and kills the TCP based chat servers.
    2. MFC has been incredibly flakey lately and having more possible servers to connect to can only be a good thing.

For now this is explicitly opt-in. No existing user scripts should be
impacted in any way.